Run golang-migrate when the server starts (behind a feature flag)
